requirements.txt
anthropic>=0.42.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pyyaml>=6.0
aiohttp>=3.9.0
requests>=2.31.0
